test(tool): move prompt protocol tests
